requirements.txt
Django==1.7b4
Pillow==2.4.0
argparse==1.2.1
django-autoslug==1.7.2
django-devserver==0.8.0
pythonwhois==2.1.3
pytils==0.3
sqlparse==0.1.11
wsgiref==0.1.2
